chore: split withErrorBoundaryGroup HOC to file with docs
