+++ packages/react/error-boundary/src/index.ts
@@ -1,3 +1,4 @@
 /** @tossdocs-ignore */
 export { ErrorBoundary, useErrorBoundary, withErrorBoundary } from './ErrorBoundary';
-export { ErrorBoundaryGroup, useErrorBoundaryGroup, withErrorBoundaryGroup } from './ErrorBoundaryGroup';
+export { ErrorBoundaryGroup, useErrorBoundaryGroup } from './ErrorBoundaryGroup';
+export { withErrorBoundaryGroup } from './withErrorBoundaryGroup';
